Making vertical_split.py always and only work with vtgate.
